// src/models/FollowingArtistAuctionPlan.d.ts
export interface FollowingArtistAuctionPlan{
    artist_id : number;
    art_info_id : number;
    auction_site : string;
    artist_name_kor : string;
    artist_name_eng : string;
    auction_cate : string;
    auction_date : string;
    auction_plan : string;
    auction_url_num : number;
    lot_no : number;
    image_name : string;
    money : number;
    image_url : string;
}